bitkeeper revision 1.272 (3f056927K6c1zv1tNa8i4njBXcuuQg)
Fix a random bunch of compiler warnings all over the tree.

+++ xenolinux-2.4.21-sparse/arch/xeno/mm/hypervisor.c
@@ -96,7 +96,7 @@ void MULTICALL_flush_page_update_queue(void)
 #if PT_UPDATE_DEBUG > 0
         DEBUG_allow_pt_reads();
 #endif
-        queue_multicall2(__HYPERVISOR_pt_update, update_queue, idx);
+        queue_multicall2(__HYPERVISOR_pt_update, (unsigned long)update_queue, idx);
         idx = 0;
     }
     spin_unlock_irqrestore(&update_lock, flags);
